gimbal: add manual control message
We realized that it often doesn't make sense to send an angular rate or
angle from ground station or remote control but really it's just a
unitless input which then gets applied by the gimbal to whatever makes
sense. For instance, the gimbal might have speed or smoothness
parameters for this, or it might take into account the zoom level of a
camera to scale its angular rate. Therefore, we propose to have unitless
input for that case.